refactor: remove account redux
